+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog for Paco
 
 ### Added
 
+- Lambda Triggers support for `CognitoUserPool` resource
+
 - Paco refs for secrets can have an additional part at the end to resolve to the JSON field inside the value of the secret.
 
 - `ECSServices` has new `setting_groups` field which allows sharing the same set of Secrets and Environments